Add animation to buttons in "ranking" status
This involves changing the class name of both ranking buttons every
time one of the buttons is clicked, and then implementing the fade-in
animation associated with the classname. This is still kind of a hacky
solution since the animations are the same logic, just differently
named, which is necessary in order to trigger the animation each time
the class name changes.